+++ commands/core.js
@@ -42,7 +42,7 @@ module.exports = [
 				.addField("Quick Stats", bot.cache.guildCount + " Servers\n" + bot.cache.userCount + ` ${userSuffix}\n` +
 					bot.cache.channelCount + " Channels", true)
 				.addField("Bot Invite",
-					`[Go!](https://discordapp.com/oauth2/authorize?client_id=${bot.user.id}&permissions=405921878&scope=bot)`, true)
+					`[Go!](https://discord.com/oauth2/authorize?client_id=${bot.user.id}&permissions=405921878&scope=bot)`, true)
 				.addField("Support Server", "[Go!](https://discord.gg/yB8TvWU)", true)
 				.addField("Upvote this bot", "discordbots.org: [Go!](https://discordbots.org/bots/333058410465722368/vote)\n" +
 					"bots.ondiscord.xyz: [Go!](https://bots.ondiscord.xyz/bots/333058410465722368)\n" +
@@ -195,7 +195,7 @@ module.exports = [
 				.setDescription("Exciting! Use these links to spread the fun!")
 				.setColor(Math.floor(Math.random() * 16777216))
 				.addField("Bot Invite",
-					`[Go!](https://discordapp.com/oauth2/authorize?client_id=${bot.user.id}&permissions=405921878&scope=bot)`, true)
+					`[Go!](https://discord.com/oauth2/authorize?client_id=${bot.user.id}&permissions=405921878&scope=bot)`, true)
 				.addField("Support Server", "[Go!](https://discord.gg/yB8TvWU)", true)
 				.addField("Upvote this bot", "discordbots.org: [Go!](https://discordbots.org/bots/333058410465722368/vote)\n" +
 					"bots.ondiscord.xyz: [Go!](https://bots.ondiscord.xyz/bots/333058410465722368)\n" +
@@ -657,7 +657,7 @@ module.exports = [
 			this.ideaWebhook.send({
 				embeds: [{
 					description: args[0].replace(/https?:\/\/\S+\.\S+/gi, "")
-						.replace(/(www\.)?(discord\.(gg|me|io)|discordapp\.com\/invite)\/[0-9a-z]+/gi, ""),
+						.replace(/(www\.)?(discord\.(gg|me|io)|discord\.com\/invite)\/[0-9a-z]+/gi, ""),
 					author: {
 						name: message.author.tag,
 						icon_url: message.author.avatarURL({format: "png", dynamic: true})